refactor(lro): move implementation to submodule
Making room for another submodule for the discovery doc LROs.
